chore: refactoring, code splitting for a better readability
