+++ src/proxy/compute/data_collection/mod.rs
@@ -1,0 +1,2 @@
+pub mod data_collection;
+pub mod components;
